requirements.txt
# Core dependencies
torch
torchvision
transformers
datasets
accelerate

# Vision and OCR
paddlepaddle
paddleocr
ultralytics
opencv-python
Pillow

# LLM and NLP
sentencepiece
protobuf
tokenizers

# Image processing
imagehash
nsfw-image-detector

# CLIP and vision models
open-clip-torch
timm

# Utilities
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
pyyaml
jsonlines

# Evaluation
pycocotools
rapidfuzz

# Development
pytest
pytest-cov
black
flake8
mypy

# Optional: For limited GPU resources
peft
bitsandbytes

# Jupyter support
jupyter
ipywidgets
ipykernel
